Cleanup folders and repositories
