+++ lib/src/features/authentication/domain/fake_app_user.dart
@@ -1,4 +1,4 @@
-import 'package:starter_architecture_flutter_firebase/src/repositories/app_user.dart';
+import 'package:starter_architecture_flutter_firebase/src/features/authentication/domain/app_user.dart';
 
 /// Fake user class used to simulate a user account on the backend
 class FakeAppUser extends AppUser {
