+++ lib/src/features/authentication/presentation/sign_in/sign_in_screen.dart
@@ -8,12 +8,7 @@ import 'package:starter_architecture_flutter_firebase/src/constants/strings.dart
 import 'package:starter_architecture_flutter_firebase/src/features/authentication/presentation/sign_in/sign_in_button.dart';
 import 'package:starter_architecture_flutter_firebase/src/features/authentication/presentation/sign_in/sign_in_view_model.dart';
 import 'package:starter_architecture_flutter_firebase/src/routing/app_router.dart';
-import 'package:starter_architecture_flutter_firebase/src/top_level_providers.dart';
 import 'package:starter_architecture_flutter_firebase/src/utils/alert_dialogs.dart';
-
-final signInModelProvider = ChangeNotifierProvider<SignInViewModel>(
-  (ref) => SignInViewModel(auth: ref.watch(firebaseAuthProvider)),
-);
 
 class SignInScreen extends ConsumerWidget {
   @override
