+++ lib/src/features/authentication/presentation/sign_in/sign_in_view_model.dart
@@ -2,6 +2,8 @@ import 'dart:async';
 
 import 'package:firebase_auth/firebase_auth.dart';
 import 'package:flutter/foundation.dart';
+import 'package:flutter_riverpod/flutter_riverpod.dart';
+import 'package:starter_architecture_flutter_firebase/src/features/authentication/data/firebase_auth_repository.dart';
 
 class SignInViewModel with ChangeNotifier {
   SignInViewModel({required this.auth});
@@ -28,3 +30,7 @@ class SignInViewModel with ChangeNotifier {
     await _signIn(auth.signInAnonymously);
   }
 }
+
+final signInModelProvider = ChangeNotifierProvider<SignInViewModel>(
+  (ref) => SignInViewModel(auth: ref.watch(firebaseAuthProvider)),
+);
