+++ lib/src/features/entries/entries_page.dart
@@ -3,8 +3,8 @@ import 'package:flutter_riverpod/flutter_riverpod.dart';
 import 'package:starter_architecture_flutter_firebase/src/constants/strings.dart';
 import 'package:starter_architecture_flutter_firebase/src/features/entries/entries_list_tile.dart';
 import 'package:starter_architecture_flutter_firebase/src/features/entries/entries_view_model.dart';
+import 'package:starter_architecture_flutter_firebase/src/features/home/data/firestore_repository.dart';
 import 'package:starter_architecture_flutter_firebase/src/features/jobs/list_items_builder.dart';
-import 'package:starter_architecture_flutter_firebase/src/top_level_providers.dart';
 
 final entriesTileModelStreamProvider =
     StreamProvider.autoDispose<List<EntriesListTileModel>>(
