+++ lib/src/features/jobs/jobs_page.dart
@@ -3,18 +3,13 @@ import 'dart:async';
 import 'package:flutter/material.dart';
 import 'package:flutter_riverpod/flutter_riverpod.dart';
 import 'package:starter_architecture_flutter_firebase/src/constants/strings.dart';
+import 'package:starter_architecture_flutter_firebase/src/features/home/data/firestore_repository.dart';
 import 'package:starter_architecture_flutter_firebase/src/features/home/models/job.dart';
 import 'package:starter_architecture_flutter_firebase/src/features/job_entries/job_entries_page.dart';
 import 'package:starter_architecture_flutter_firebase/src/features/jobs/edit_job_page.dart';
 import 'package:starter_architecture_flutter_firebase/src/features/jobs/job_list_tile.dart';
 import 'package:starter_architecture_flutter_firebase/src/features/jobs/list_items_builder.dart';
-import 'package:starter_architecture_flutter_firebase/src/top_level_providers.dart';
 import 'package:starter_architecture_flutter_firebase/src/utils/alert_dialogs.dart';
-
-final jobsStreamProvider = StreamProvider.autoDispose<List<Job>>((ref) {
-  final database = ref.watch(databaseProvider);
-  return database.jobsStream();
-});
 
 // watch database
 class JobsPage extends ConsumerWidget {
