--- lib/src/top_level_providers.dart
@@ -1,26 +1,0 @@
-import 'package:firebase_auth/firebase_auth.dart';
-import 'package:flutter_riverpod/flutter_riverpod.dart';
-import 'package:logger/logger.dart';
-import 'package:starter_architecture_flutter_firebase/src/services/firestore_database.dart';
-
-final firebaseAuthProvider =
-    Provider<FirebaseAuth>((ref) => FirebaseAuth.instance);
-
-final authStateChangesProvider = StreamProvider<User?>(
-    (ref) => ref.watch(firebaseAuthProvider).authStateChanges());
-
-final databaseProvider = Provider<FirestoreDatabase>((ref) {
-  final authStateAsync = ref.watch(authStateChangesProvider);
-
-  if (authStateAsync.value?.uid != null) {
-    return FirestoreDatabase(uid: authStateAsync.value!.uid);
-  }
-  throw UnimplementedError();
-});
-
-final loggerProvider = Provider<Logger>((ref) => Logger(
-      printer: PrettyPrinter(
-        methodCount: 1,
-        printEmojis: false,
-      ),
-    ));
