+++ lib/src/utils/logger_provider.dart
@@ -1,0 +1,9 @@
+import 'package:flutter_riverpod/flutter_riverpod.dart';
+import 'package:logger/logger.dart';
+
+final loggerProvider = Provider<Logger>((ref) => Logger(
+      printer: PrettyPrinter(
+        methodCount: 1,
+        printEmojis: false,
+      ),
+    ));
